Make `Match` hierarchically in a tree structure
Moved the position of the nested `Match` components. To support proper
styling, more specific CSS classes have been introduced. Additionally,
because it may be convenient to know where each element should belong,
`data-position` has been added. That should make it easier to identify
specific elements.

Sadly Happy-DOM has an issue with direct `>` selectors, so it's very
hard to create unit tests for a recursive structure as this. That's
because all nested elements are also found. We could move things around
a bit in the template to make sure `.match-details` is first, but that
defeats the purpose a bit. This is why the tests are marked `todo` and
https://github.com/capricorn86/happy-dom/issues/564 has been logged.