Fix re-rendering problem with nested `Match` components
Only when Vue3 decided to completely re-render the `Match` component,
any added child `Match` components would be rendered. But because that's
not always the case, there was an issue, in particular when a `Match`
went from two direct values, to having a child `Match`.

The cause is that `hasMatchModel` was stored only when the `Match`
component was rendered, but the child `Match` components are dependent
on it. So now changed it to a `computed`, which tells Vue3 to properly
render the child `Match`.